+++ sdks/java/io/snowflake/src/main/java/org/apache/beam/sdk/io/snowflake/services/SnowflakeBatchServiceConfig.java
@@ -24,79 +24,118 @@ import org.apache.beam.sdk.io.snowflake.enums.CreateDisposition;
 import org.apache.beam.sdk.io.snowflake.enums.WriteDisposition;
 import org.apache.beam.sdk.transforms.SerializableFunction;
 
-public class SnowflakeServiceConfig extends ServiceConfig {
-  private SerializableFunction<Void, DataSource> dataSourceProviderFn;
+/** Class for preparing configuration for batch write and read. */
+public class SnowflakeBatchServiceConfig extends ServiceConfig {
+  private final SerializableFunction<Void, DataSource> dataSourceProviderFn;
 
-  private String table;
-  private String query;
-  private String storageIntegrationName;
+  private final String database;
+  private final String schema;
+  private final String table;
+  private final String query;
+  private final String storageIntegrationName;
   private List<String> filesList;
-
   private WriteDisposition writeDisposition;
   private CreateDisposition createDisposition;
   private SnowflakeTableSchema tableSchema;
-  private String stagingBucketDir;
+  private final String stagingBucketDir;
+  private final String quotationMark;
 
-  public SnowflakeServiceConfig(
+  /** Creating a batch configuration for reading. */
+  public SnowflakeBatchServiceConfig(
       SerializableFunction<Void, DataSource> dataSourceProviderFn,
+      String database,
+      String schema,
       String table,
       String query,
-      String storageIntegration,
-      String stagingBucketDir) {
+      String storageIntegrationName,
+      String stagingBucketDir,
+      String quotationMark) {
     this.dataSourceProviderFn = dataSourceProviderFn;
+    this.database = database;
+    this.schema = schema;
     this.table = table;
     this.query = query;
-    this.storageIntegrationName = storageIntegration;
+    this.storageIntegrationName = storageIntegrationName;
     this.stagingBucketDir = stagingBucketDir;
+    this.quotationMark = quotationMark;
   }
 
-  public SnowflakeServiceConfig(
+  /** Creating a batch configuration for writing. */
+  public SnowflakeBatchServiceConfig(
       SerializableFunction<Void, DataSource> dataSourceProviderFn,
       List<String> filesList,
+      String database,
+      String schema,
       String table,
       String query,
       SnowflakeTableSchema tableSchema,
       CreateDisposition createDisposition,
       WriteDisposition writeDisposition,
       String storageIntegrationName,
-      String stagingBucketDir) {
+      String stagingBucketDir,
+      String quotationMark) {
     this.dataSourceProviderFn = dataSourceProviderFn;
     this.filesList = filesList;
+    this.database = database;
+    this.schema = schema;
     this.table = table;
     this.query = query;
-    this.tableSchema = tableSchema;
     this.writeDisposition = writeDisposition;
     this.createDisposition = createDisposition;
+    this.tableSchema = tableSchema;
     this.storageIntegrationName = storageIntegrationName;
     this.stagingBucketDir = stagingBucketDir;
+    this.quotationMark = quotationMark;
   }
 
+  /** Getting a DataSource provider function for connection credentials. */
   public SerializableFunction<Void, DataSource> getDataSourceProviderFn() {
     return dataSourceProviderFn;
   }
 
+  /** Getting a table as a source of reading or destination to write. */
   public String getTable() {
     return table;
   }
 
+  /** Getting a query which can be source for reading. */
   public String getQuery() {
     return query;
   }
 
+  /** Getting Snowflake integration which is used in COPY statement. */
   public String getStorageIntegrationName() {
     return storageIntegrationName;
   }
 
+  /** Getting directory where files are staged. */
   public String getStagingBucketDir() {
     return stagingBucketDir;
   }
 
+  /** Getting list of names of staged files. */
   public List<String> getFilesList() {
     return filesList;
   }
 
+  /** Getting disposition how write data to table, see: {@link WriteDisposition}. */
   public WriteDisposition getWriteDisposition() {
     return writeDisposition;
+  }
+
+  /** Getting a character that will surround {@code String} in staged CSV files. */
+  public String getQuotationMark() {
+    return quotationMark;
+  }
+
+  /** Getting a Snowflake database. */
+  public String getDatabase() {
+    return database;
+  }
+
+  /** Getting a schema of a Snowflake table. */
+  public String getSchema() {
+    return schema;
   }
 
   public CreateDisposition getCreateDisposition() {